Refactor player options and repeat hook
